requirements.txt
fastapi
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
mysqlclient
python-dotenv
sqlalchemy
uvicorn[standard]
